+++ app/feature/splitter/Splitter.tsx
@@ -4,7 +4,7 @@ import React, { lazy, Suspense } from "react";
 const Split = lazy(() => import("react-split"));
 import QuestionSection from "@/app/components/sections/QuestionSection";
 import CodeInputSection from "@/app/feature/monacoEditor/sections/CodeInputSection";
-import ReviewSection from "@/app/components/layout/reviewSection/ReviewSection";
+import ReviewSection from "@/app/components/sections/ReviewSection";
 import { usePanelContext } from "@/app/feature/splitter/context/PanelContext";
 
 // 垂直方向にパネルを分割するコンポーネント
