+++ app/type/type.ts
@@ -204,8 +204,8 @@ export interface ProblemContentProps extends ProblemContentBase {
   hints?: string;
 }
 
-// ReusableParagraph(APIからのレスポンスを展開する)に対する型定義
-export interface ReusableProblemContentProps {
+// Paragraph(APIからのレスポンスを展開する)に対する型定義
+export interface QuestionParagraphProps {
   content: ProblemContentProps | null;
   titleText: string;
   paragraphContent: string | null | undefined;
@@ -222,8 +222,8 @@ export interface ReviewResponseProps {
   improvementExample: string | null;
 }
 
-// ReusableParagraph(APIからのレスポンスを展開する)に対する型定義
-export interface ReusableReviewContentsProps {
+// Paragraph(APIからのレスポンスを展開する)に対する型定義
+export interface ReviewParagraphProps {
   content: ReviewResponseProps | null;
   titleText: string;
   paragraphContent: string | null | undefined;
